made changes to message protocol to handle int16 return types

+++ include/message.h
@@ -15,7 +15,6 @@
 #define ACK                 0x06U
 #define NAK                 0x15U
 
-
 #define ERROR_INVALID_PREAMBLE  -1
 #define ERROR_FSM_FAULT         -2
 
@@ -46,7 +45,7 @@ enum messgage_id {
 
 
 
-typedef int8_t (*handler_func_t)(void *data);
+typedef int16_t (*handler_func_t)(void *data);
 extern handler_func_t handler_table[HANDLERS_TABLE_SZ];
 
 enum sc_state {
@@ -56,8 +55,8 @@ enum sc_state {
 	STOP
 };
 
-int8_t process_message(struct ringbuffer *buff);
-int8_t dispatch(struct message *msg, handler_func_t table[]);
+int16_t process_message(struct ringbuffer *buff);
+int16_t dispatch(struct message *msg, handler_func_t table[]);
 void print_message(struct message *msg);
 
 //void wrap_message(uint8_t id, const uint8_t *data);
